refactor: move compare, evaluate_sample, and helpers to TokenEvaluator
compare(), __revert_known_errors(), _adjust_per_entities(), and
evaluate_sample() are token-level concerns only used by TokenEvaluator.
Move them out of BaseEvaluator and into TokenEvaluator where they belong.

- BaseEvaluator: remove compare, __revert_known_errors, _adjust_per_entities,
  evaluate_sample; inline _adjust_per_entities into deprecated get_results_dataframe;
  remove now-unused spacy Token import
- TokenEvaluator: add all four methods; add logging + spacy Token + ErrorType + ModelError imports
- MockEvaluator in test_evaluator.py: inherit from TokenEvaluator (was BaseEvaluator)
- Fix stale Evaluator(, ...) patterns across test files and docs (add model=None)
